Update kaprekarUtils.c
Got rid of (int).

+++ c_version/kaprekarUtils.c
@@ -32,7 +32,7 @@ int isKaprekar(int n) {
   //compute the square
   long square = n * n;
   //compute the number of digits of the square
-  int numDigits = (int) log10(n) + 1;
+  int numDigits = log10(n) + 1;
   long modulus = 0;
   long first, second;
   double numberOfDigits, modulous;
